src/main/java/com/blank04/universitycms/controller/entitycontoller/GroupController.java
package com.blank04.universitycms.controller.entitycontoller;

import com.blank04.universitycms.model.entity.Group;
import com.blank04.universitycms.service.GroupService;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;

import java.util.List;

@Controller
@RequestMapping("/api/groups")
public class GroupController {

    private final GroupService groupService;

    public GroupController(GroupService groupService) {
        this.groupService = groupService;
    }

    @GetMapping
    public String findAll(Model model) {
        List<Group> groups = groupService.findAll();
        model.addAttribute("groups", groups);
        return "groups-list";
    }
}
